Added users/groups, ROI conversions and migrated to Github Actions (#10)
* Added users/groups classes and ROI functions.

* Added missing tests for users/groups and removed non-functional methods for now.

* Added method to export OMERO ROIs to ImageJ.

* Modified method to export OMERO ROIs to ImageJ.

* Small adjustments to order objects lists.

* Added method to import ImageJ ROIs to OMERO.

* Switched from Tracis-CI to Github Actions.

* Added method to switch group and improved "toImageJ" method for some rotated shapes.

* Fixed test.

* Improved ROI conversion and added tests.

* Added artifacts to Github Actions (#9)

* Changed "setCoordinates" methods in shapes.

* Reverted some changes (to increase coverage).

* Removed shallow clone, for sonarcloud.

* Changed CZT conversion to ImageJ.

+++ src/test/java/fr/igred/omero/annotations/TagTest.java
@@ -31,7 +31,7 @@ public class TagTest extends UserTest {
     @Test
     public void testGetTagInfo() throws Exception {
         TagAnnotationWrapper tag = client.getTag(1L);
-        assertEquals(1L, tag.getId().longValue());
+        assertEquals(1L, tag.getId());
         assertEquals("tag1", tag.getName());
         assertEquals("description", tag.getDescription());
     }
